Refactor tests
Move code to separate files
