+++ test-app/spec/_spec_helper.rb
@@ -1,3 +1,5 @@
+require "eapol_test_helper"
+
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # The generated `.rspec` file contains `--require spec_helper` which will cause
@@ -16,7 +18,17 @@
 # users commonly want.
 #
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
+#
+require "logging_line"
+require "auth_line"
+
 RSpec.configure do |config|
+  config.before(:each) do
+    LoggingLine.truncate
+    AuthLine.truncate
+  end
+
+  config.include EapolTestHelper
   # rspec-expectations config goes here. You can use an alternate
   # assertion/expectation library such as wrong or the stdlib/minitest
   # assertions if you prefer.
