+++ test-app/spec/auth_line.rb
@@ -1,0 +1,4 @@
+DB_AUTH = Sequel.connect(adapter: 'sqlite', database: ENV.fetch("AUTH_DB"))
+
+class AuthLine < Sequel::Model(DB_AUTH[:lines])
+end
