+++ test-app/spec/auth_spec.rb
@@ -2,17 +2,9 @@ require "sinatra/base"
 require "json"
 require 'sequel'
 require 'sqlite3'
-
 require "commands"
-
-DB_AUTH = Sequel.connect(adapter: 'sqlite', database: ENV.fetch("AUTH_DB"))
-DB_LOGGING = Sequel.connect(adapter: 'sqlite', database: ENV.fetch("LOGGING_DB"))
-
-class AuthLine < Sequel::Model(DB_AUTH[:lines])
-end
-
-class LoggingLine < Sequel::Model(DB_LOGGING[:lines])
-end
+require "eapol_test_helper"
+require "_spec_helper"
 
 RSpec.shared_examples "it rejects authentication attempt" do |command|
   let(:command) { command }
@@ -24,27 +16,7 @@ RSpec.shared_examples "it rejects authentication attempt" do |command|
   end
 end
 
-def run_eapol(config_template_path, variables={})
-  variables_with_defaults = { phase1: "",
-                              phase2: "",
-                              username: ENV.fetch("HEALTH_CHECK_IDENTITY"),
-                              password: ENV.fetch("HEALTH_CHECK_PASSWORD") }.merge(variables)
-  file = Tempfile.new
-  file.write ERB.new(File.read(config_template_path)).result_with_hash(variables_with_defaults)
-  file.close
-
-  command = "eapol_test -a 127.0.0.1 -c #{file.path} -s testing123"
-  `#{command}`
-ensure
-  file.unlink
-end
-
 RSpec.describe 'test' do
-  before :each do
-    LoggingLine.truncate
-    AuthLine.truncate
-  end
-
   it_behaves_like "it rejects authentication attempt", PAP_CMD
   it_behaves_like "it rejects authentication attempt", CHAP_CMD
   it_behaves_like "it rejects authentication attempt", MSCHAP_CMD
